Allow bigfloats through dualcaches
We could in theory embed a LBC into the DualCache so that it reuses this vector, but I'm not sure it's worth it. Basically, if you have a T that's not bitstype then it's not going to reuse memory and it's going to have to re-allocate every number, so you have lots of memory operations already as requirements for making the T being acted on. In this case, you're really just looking for a functional fallback as there is no such thing as preallocation, so simply making the array you wish to see is a good option.